Add unit tests for applicationState
